refactor: move secrets provider setup to root
